Move definition for api factory
